Wire into ksp
